green/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
clickup-python==0.1.1
python-dotenv==1.0.1 